refactor: Simplify repository
