+++ .flake8
@@ -2,4 +2,5 @@
 max-line-length = 88
 extend-ignore = E203
 max-complexity = 18
-
+per-file-ignores =
+    **/__init__.py:F401
