--- src/coral_models/asr/__init__.py
@@ -1,4 +1,0 @@
-"""Modules related to the finetuning models for Automatic Speech Recognition (ASR)."""
-
-from .ftspeech import build_and_store_ftspeech  # noqa: F401
-from .wav2vec2 import finetune_wav2vec2  # noqa: F401
