--- src/coral_models/asr/utils.py
@@ -1,31 +1,0 @@
-"""Utility functions to be used in other ASR scripts."""
-
-import json
-from pathlib import Path
-
-from omegaconf import DictConfig
-
-
-def dump_vocabulary(cfg: DictConfig) -> None:
-    """Extracts the vocabulary from the dataset and dumps it to a file.
-
-    It will dump the file to `${cfg.model_dir}/vocab.json`.
-
-    Args:
-        cfg (DictConfig):
-            The Hydra configuration object.
-    """
-    # Build the set of all unique characters in the dataset
-    unique_characters: set[str] = set(cfg.characters_to_keep)
-
-    # Build vocabulary
-    vocab = {char: idx for idx, char in enumerate(unique_characters)}
-    for tok in ["<unk>", "<pad>", "<s>", "</s>"]:
-        vocab[tok] = len(vocab)
-
-    # Dump the vocabulary to a json file
-    model_dir = Path(cfg.model_dir)
-    model_dir.mkdir(parents=True, exist_ok=True)
-    vocab_path = model_dir / "vocab.json"
-    with vocab_path.open("w") as f:
-        json.dump(vocab, f)
