--- src/coral_models/asr/wav2vec2/__init__.py
@@ -1,3 +1,0 @@
-"""Modules related to the finetuning Wav2Vec 2.0 models on ASR datasets."""
-
-from .finetune import finetune_wav2vec2
