--- src/coral_models/asr/wav2vec2/preprocess.py
@@ -1,124 +1,0 @@
-"""Functions related to the preprocessing of the data for Wav2Vec 2.0 models."""
-
-from datasets import Audio, DatasetDict, IterableDatasetDict
-from omegaconf import DictConfig
-from transformers import (
-    PreTrainedTokenizerBase,
-    Wav2Vec2CTCTokenizer,
-    Wav2Vec2FeatureExtractor,
-    Wav2Vec2Processor,
-)
-
-
-class ModifiedWav2Vec2Processor(Wav2Vec2Processor):
-    """The preprocessor for the Wav2Vec 2.0 model.
-
-    This class extends the `Wav2Vec2Processor` class from the Hugging Face
-    `transformers` package, adding the `preprocess` method to preprocess the audio and
-    transcriptions in a dataset.
-
-    Args:
-        cfg (DictConfig):
-            The Hydra configuration object.
-    """
-
-    def __init__(self, cfg: DictConfig, **kwargs) -> None:
-        self.cfg = cfg
-        self.tokenizer: PreTrainedTokenizerBase = kwargs["tokenizer"]
-        super().__init__(**kwargs)
-
-    def preprocess(
-        self, dataset: DatasetDict | IterableDatasetDict
-    ) -> DatasetDict | IterableDatasetDict:
-        """Preprocess the audio and transcriptions in a dataset.
-
-        Args:
-            dataset (DatasetDict or IterableDatasetDict):
-                The dataset containing the audio and transcriptions.
-
-        Returns:
-            DatasetDict or IterableDatasetDict:
-                The preprocessed audio and transcriptions dataset.
-        """
-        dataset = self.preprocess_audio(dataset=dataset)
-        dataset = self.preprocess_transcriptions(dataset=dataset)
-        return dataset
-
-    def preprocess_audio(
-        self, dataset: DatasetDict | IterableDatasetDict
-    ) -> DatasetDict | IterableDatasetDict:
-        """Preprocess the audio feature in a dataset.
-
-        Args:
-            dataset (DatasetDict or IterableDatasetDict):
-                The dataset containing the audio feature.
-
-        Returns:
-            DatasetDict or IterableDatasetDict:
-                The preprocessed audio dataset.
-        """
-        audio = Audio(sampling_rate=self.cfg.model.sampling_rate)
-        return dataset.cast_column("audio", audio)
-
-    def preprocess_transcriptions(
-        self, dataset: DatasetDict | IterableDatasetDict
-    ) -> DatasetDict | IterableDatasetDict:
-        """Preprocess the transcriptions in a dataset.
-
-        Args:
-            dataset (DatasetDict or IterableDatasetDict):
-                The dataset containing the transcriptions.
-
-        Returns:
-            DatasetDict or IterableDatasetDict:
-                The preprocessed transcriptions dataset.
-        """
-
-        def tokenize_examples(example: dict) -> dict:
-            example["labels"] = self(
-                text=example[self.cfg.dataset.text_column],
-                truncation=True,
-            ).input_ids
-            example["input_length"] = len(example["labels"])
-            return example
-
-        return dataset.map(tokenize_examples)
-
-
-def load_processor(cfg: DictConfig) -> ModifiedWav2Vec2Processor:
-    """Load the processor for a Wav2Vec 2.0 model.
-
-    Args:
-        cfg (DictConfig):
-            The Hydra configuration object.
-
-    Returns:
-        ModifiedWav2Vec2Processor:
-            The processor for a Wav2Vec 2.0 model.
-    """
-    tokenizer: Wav2Vec2CTCTokenizer = Wav2Vec2CTCTokenizer.from_pretrained(
-        cfg.model_dir,
-        unk_token="<unk>",
-        pad_token="<pad>",
-        bos_token="<s>",
-        eos_token="</s>",
-        word_delimiter_token="|",
-    )
-
-    # Set the `model_max_length` attribute of the tokenizer, if it hasn't been set.
-    # This property defaults to very large value if not set, which causes issues with
-    # truncation, so we check manually if it is above a million, as this is unlikely to
-    # be a valid value.
-    if tokenizer.model_max_length is None or tokenizer.model_max_length > 1e6:
-        tokenizer.model_max_length = 512
-
-    extractor = Wav2Vec2FeatureExtractor(
-        feature_size=1,
-        sampling_rate=cfg.model.sampling_rate,
-        padding_value=0.0,
-        do_normalize=True,
-        return_attention_mask=True,
-    )
-    return ModifiedWav2Vec2Processor(
-        cfg, feature_extractor=extractor, tokenizer=tokenizer
-    )
