--- src/coral_models/asr/whisper/__init__.py
@@ -1,1 +1,0 @@
-"""Modules related to the finetuning Whisper models on ASR datasets."""
