--- src/coral_models/tts/__init__.py
@@ -1,1 +1,0 @@
-"""Modules related to the finetuning models for Text-To-Speech (TTS)."""
